Small fix to Dynamic PTI
At fork(), when a process is created, the mm_struct is created as a mere
copy of the parent's one. The flags are copied and cleaned up with
MMF_INIT_MASK. This action creates a conflict when the process is
sharing the PGD, but holds its copy such that the flags are not
compliant with the PGD state.

This commit introduces a primary fix that extends the flags to all those
processes forked by a PTI-enabled process. The final goal of this set
of patches is to avoid enabling PTI for forked processes which will run
in an unrelated context from the parent.