Add virtual PMC CPL filtering
This commit introduces a preliminary implementation of a
virtual CPL filtering for PMC. This support is already present in HW
but is known to be buggy.

We patched the entry and exit points of the kernel such that, through a
set of snapshots which are compared at each transition (K->U) and then
the USR activity is computed. At each context_switch is computed
the OS activity for the exiting process.

Note that the ESPFIX entry is not patched in entry_64.S because it
should be just a fix for "to-16-bit transition" and the amount of code
is negligible as well as it does not really contribute to USR action
which is, in the end, totally attributed to OS.

Even though there is a small support dynamic CPL filter, we need to
build more flexible support according to the number of PMC that
the underlying processor contains.